projects/chapter-07-project/requirements.txt
# Chapter 7 Project Requirements

# Google Gemini API
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0
